# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Objective Index",
    "description": "Read-only objective-fit ranking API for local AOI sample records.",
    "version": "0.1"
  },
  "paths": {
...
    "/search": {
      "get": {
        "summary": "Search",
        "operationId": "search_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          },
          {
            "name": "objective",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Objective"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "data_scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(sample|generated|integrated|curated|public_beta|mcp_registry|public_beta_mcp)$",
              "default": "sample",
              "title": "Data Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/objects/{object_id}": {
      "get": {
        "summary": "Get Object",
        "operationId": "get_object_objects__object_id__get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "data_scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(sample|generated|integrated|curated|public_beta|mcp_registry|public_beta_mcp)$",
              "default": "sample",
              "title": "Data Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rank": {
      "post": {
        "summary": "Rank",
        "operationId": "rank_rank_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Rank Rank Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/compare": {
      "post": {
        "summary": "Compare",
        "operationId": "compare_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Compare Compare Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/objects/{object_id}/score": {
      "get": {
        "summary": "Score",
        "operationId": "score_objects__object_id__score_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "data_scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(sample|generated|integrated|curated|public_beta|mcp_registry|public_beta_mcp)$",
              "default": "sample",
              "title": "Data Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/objects/{object_id}/source-trace": {
      "get": {
        "summary": "Source Trace",
        "operationId": "source_trace_objects__object_id__source_trace_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "field",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Field"
            }
          },
          {
            "name": "data_scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(sample|generated|integrated|curated|public_beta|mcp_registry|public_beta_mcp)$",
              "default": "sample",
              "title": "Data Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/objects/{object_id}/missing-fields": {
      "get": {
        "summary": "Missing Fields",
        "operationId": "missing_fields_objects__object_id__missing_fields_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "data_scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(sample|generated|integrated|curated|public_beta|mcp_registry|public_beta_mcp)$",
              "default": "sample",
              "title": "Data Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/decision-receipt": {
      "post": {
        "summary": "Decision Receipt",
        "operationId": "decision_receipt_decision_receipt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionReceiptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CompareRequest": {
        "properties": {
          "object_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Object Ids"
          },
          "compare_fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compare Fields"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "objective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Constraints"
          },
          "data_scope": {
            "type": "string",
            "title": "Data Scope",
            "default": "sample"
          }
        },
        "type": "object",
        "required": [
          "object_ids"
        ],
        "title": "CompareRequest"
      },
      "DecisionReceiptRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "selected_object_id": {
            "type": "string",
            "title": "Selected Object Id"
          },
          "alternatives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alternatives"
          },
          "objective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Constraints"
          },
          "data_scope": {
            "type": "string",
            "title": "Data Scope",
            "default": "sample"
          }
        },
        "type": "object",
        "required": [
          "query",
          "selected_object_id"
        ],
        "title": "DecisionReceiptRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RankOption": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RankOption"
      },
      "RankOptionsRequest": {
        "properties": {
          "options": {
            "items": {
              "$ref": "#/components/schemas/RankOption"
            },
            "type": "array",
            "title": "Options"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "constraints": {
            "additionalProperties": true,
            "type": "object",
            "title": "Constraints"
          },
          "scoring_profile": {
            "type": "string",
            "title": "Scoring Profile",
            "default": "default"
          },
          "data_scope": {
            "type": "string",
            "title": "Data Scope",
            "default": "sample"
          }
        },
        "type": "object",
        "required": [
          "options",
          "objective"
        ],
        "title": "RankOptionsRequest"
      },
      "SearchResponse": {
        "properties": {
          "read_only": {
            "type": "boolean",
            "title": "Read Only",
            "default": true
          },
          "data_scope": {
            "type": "string",
            "title": "Data Scope",
            "default": "sample"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "objective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Results"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "forbidden_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Forbidden Actions"
          },
          "action_boundary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Action Boundary"
          }
        },
        "type": "object",
        "required": [
          "query",
          "results",
          "limitations",
          "forbidden_actions"
        ],
        "title": "SearchResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}